+++ CHANGES.md
@@ -37,6 +37,7 @@ Change Log
 * Fixed bug causing `ArcGisFeatureServerCatalogItem` to throw an error when a token is included in the proxy url.
 * Fix a bug for zooming to `ArcGisMapServerCatalogItem` layers
 * Modified creation of catalog item from urls to set the item name to be the url at the defaults stratum rather than the definition stratum. This prevents actual item names at load strata from being overridden by a definition stratum name which is just a url.
+* Fixed bug causing user added data with an auto-detected data type to not be shared properly.
 * Modified `addToWorkbench` so that when a catalog item fails to load it is removed from the workbench and an error message is displayed.
 * Add support for feature picking on region mapped datasets
 * Revamp map buttons at top to support two menu configuration
@@ -46,6 +47,7 @@ Change Log
 * Added AccessControlMixin for tracking access control of a given MagdaReference
 * Add a legend title trait
 * Show private or public dataset status on data catalog UI via AccessControlMixin
+* Added `pointSizeMap` to `TableStyle` to allow point size to be scaled by value
 
 ### Next Release
 * Fix draggable workbench/story items with translation HOC